Ai_Assistant/Qwen-openrouter/requirements.txt
flask
openai

python-dotenv
# Recommended for production deployment
gunicorn